+++ radiate/src/models/neat/layers/mod.rs
@@ -19,7 +19,7 @@ pub mod vectorops {
 
         /// multiply two vectors element wise
         #[inline]
-        pub fn element_multiply(one: &mut Vec<f64>, two: &Vec<f64>) {
+        pub fn element_multiply(one: &mut Vec<f32>, two: &Vec<f32>) {
             one.iter_mut()
                 .zip(two.iter())
                 .for_each(|(a, b)| {
@@ -31,7 +31,7 @@ pub mod vectorops {
     
         /// invert a vector that is already holding values between 0 and 1
         #[inline]
-        pub fn element_invert(one: &mut Vec<f64>) {
+        pub fn element_invert(one: &mut Vec<f32>) {
             one.iter_mut()
                 .for_each(|a| *a = 1.0 - *a);
         }
@@ -40,7 +40,7 @@ pub mod vectorops {
     
         /// add elements from vectors together element wise
         #[inline]
-        pub fn element_add(one: &mut Vec<f64>, two: &Vec<f64>) {
+        pub fn element_add(one: &mut Vec<f32>, two: &Vec<f32>) {
             one.iter_mut()
                 .zip(two.iter())
                 .for_each(|(a, b)| {
